requirements.txt
fastapi
httpx
pydantic
python-dotenv
uvicorn
pyjwt
pydantic_settings